#1 feat: slide 수정 및 스크롤 이벤트 수정
