+++ src/components/MyCardBtn.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useMutation } from "react-query";
 import { useNavigate } from "react-router-dom";
-import { useRecoilState } from "recoil";
+import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
 import styled from "styled-components";
 import Swal from "sweetalert2";
 import { memberListuseMutationGetCardInfo } from "../apis/queries/memberListQuery";
@@ -12,6 +12,7 @@ import {
   instagram,
   introduceText,
   language,
+  nicknameSelector,
   nickNameText,
   position,
   skillSrc,
@@ -22,6 +23,7 @@ import {
 
 const MyCardBtn = ({ setClick }) => {
   const navigate = useNavigate();
+  // const setStatusMsg = useSetRecoilState(statusMessageSelector);
 
   const [statusMsg, setStatusMsg] = useRecoilState(statusMessage);
   const [field, setField] = useRecoilState(position);
@@ -41,6 +43,7 @@ const MyCardBtn = ({ setClick }) => {
     {
       onSuccess: (res) => {
         console.log(res);
+        setNickName(res.nickname);
         setField(res.field);
         setIntroduce(res.introduction);
         setSkill(res.skill);
@@ -48,7 +51,6 @@ const MyCardBtn = ({ setClick }) => {
         setTagList(res.tagSet);
         setImgUrl(res.profileImage);
         setSkillUrl(res.skillImage);
-        setNickName(localStorage.getItem("name"));
         res.snsSet.map((data) => {
           if (data.name === "twitter") return setTwitterId(data.account);
           if (data.name === "github") return setGithubId(data.account);
@@ -68,15 +70,9 @@ const MyCardBtn = ({ setClick }) => {
         confirmButtonColor: "#DCC6C6",
         cancelButtonColor: "#738598",
         showCancelButton: true,
-        confirmButtonText: "로그인하기",
-        cancelButtonText: "NO",
+        confirmButtonText: "로그인",
+        cancelButtonText: "취소",
         padding: "2em",
-        showClass: {
-          popup: "animate__animated animate__fadeInDown",
-        },
-        hideClass: {
-          popup: "animate__animated animate__fadeOutUp",
-        },
         closeOnClickOutside: false,
       }).then((data) => {
         if (data.isConfirmed) navigate("/login");
