+++ src/store/userInfoAtom.js
@@ -46,53 +46,3 @@ export const loginGithubId = atom({
   default: "",
 })
 
-
-// export const userNicknameSelector = selector({
-//   key: `user/${v1()}`,
-//   get: ({ get }) => {
-//     const nickname = get(userNickname);
-//     return nickname;
-//   },
-//   set: ({ set }, newValue) => {
-//     console.log(newValue);
-//     set(userNickname, newValue);
-//   },
-// });
-
-export const userEmailSelector = selector({
-  key: `user/${v1()}`,
-    get: async ({ get }) => {
-      const res = await cookieApi.get("/user/valid");
-      return res.data.email;
-    },
-    set: ({ set }, newValue) => {
-      console.log(newValue);
-      set(loginUserEmail, newValue);
-  }
-});
-
-export const userNicknameSelector = selector({
-  key: `user/${v1()}`,
-  get: ({ get }) => {
-    return get(loginUserNickname);
-  },
-});
-
-export const userPasswordLengthSelector = selector({
-  key: `user/${v1()}`,
-  get: ({ get }) => {
-    return get(loginUserPasswordLength);
-  },
-});
-
-// export const userPasswordSelector = selector({
-//   key: `user/${v1()}`,
-//   get: ({ get }) => {
-//     const password = get(userPassword);
-//     return password;
-//   },
-//   set: ({ set }, newValue) => {
-//     console.log(newValue);
-//     set(userPassword, newValue);
-//   },
-// });
