+++ src/style/GlobalStyle.js
@@ -21,7 +21,10 @@ const GlobalStyle = createGlobalStyle`
   body {
     margin: 0;
     padding: 0;
-    /* overflow: hidden; */
+    #root{
+        /* width: 100%;
+        height: 100%; */
+    }
   }
   p {
     margin:0;
